refactor(ui): Rename components
Bulk curations are now called bulk conclusions. This commit renames
components and the corresponding imports to reflect this change.
