+++ apps/clearance_ui/src/components/delete_item/DeleteBulkConclusion.tsx
